@@ -1,9 +1,8 @@
 package dev.ronin.demo.beerstore.domain.customer.data;
 
 import dev.ronin.demo.beerstore.domain.customer.value.Address;
+import jakarta.persistence.*;
 import lombok.*;
-
-import javax.persistence.*;
 
 @Getter
 @Setter
@@ -12,6 +11,12 @@ import javax.persistence.*;
 @NoArgsConstructor
 @Entity(name = "CUSTOMER")
 public class CustomerData {
+
+    public CustomerData(String firstName, String lastName, Address address) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.address = address;
+    }
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
