@@ -1,18 +1,13 @@
 package dev.ronin.demo.beerstore.infrastructure.controller;
 
-import dev.ronin.demo.beerstore.contract.customerdata.CustomerModel;
 import dev.ronin.demo.beerstore.infrastructure.adapter.CustomersAdapter;
+import dev.ronin.demo.beerstore.infrastructure.api.CustomerApi;
+import dev.ronin.demo.beerstore.infrastructure.api.model.CustomerModel;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-
 @RestController
-@RequestMapping(value = "/customers", produces = APPLICATION_JSON_VALUE)
-public class CustomerController {
+public class CustomerController implements CustomerApi {
 
     final CustomersAdapter customersAdapter;
 
@@ -20,8 +15,8 @@ public class CustomerController {
         this.customersAdapter = customersAdapter;
     }
 
-    @GetMapping(value = "/{name}")
-    public ResponseEntity<CustomerModel> getOrderById(@PathVariable final String name) {
+    @Override
+    public ResponseEntity<CustomerModel> getCustomerByName(String name) {
         return ResponseEntity.ok(customersAdapter.customerWithName(name));
     }
 }
